moved selectedTrip into ViewModel, reorganized files, refactored code
