+++ app/src/main/java/com/example/tripline/fragments/SearchFragment.java
@@ -1,4 +1,4 @@
-package com.example.tripline.ui.search;
+package com.example.tripline.fragments;
 
 import android.os.Bundle;
