+++ app/src/main/res/layout/fragment_follower.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FollowerFragment">
+    tools:context=".fragments.FollowerFragment">
 
     <TextView
         android:id="@+id/tvFollowersTitle"
