+++ app/src/main/res/layout/fragment_following.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FollowingFragment">
+    tools:context=".fragments.FollowingFragment">
 
     <TextView
         android:id="@+id/tvFollowingTitle"
