+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".ui.search.SearchFragment">
+    tools:context=".fragments.SearchFragment">
 
 
     <SearchView
