+++ app/src/main/res/layout/item_event.xml
@@ -41,18 +41,17 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="4dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         android:layout_marginEnd="4dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/ivIconEvent" />
+        app:layout_constraintTop_toBottomOf="@+id/tvDescriptionEvent" />
 
     <TextView
         android:id="@+id/tvDescriptionEvent"
-        android:layout_width="344dp"
+        android:layout_width="296dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        app:layout_constraintBottom_toBottomOf="@+id/ivIconEvent"
+        android:layout_marginStart="12dp"
         app:layout_constraintStart_toEndOf="@+id/ivIconEvent"
         app:layout_constraintTop_toBottomOf="@+id/tvTripTitleEvent"
         tools:text="a super cool snorkeling trip yay" />
