+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_stream"
-        android:name="com.example.tripline.ui.stream.StreamFragment"
+        android:name="com.example.tripline.fragments.StreamFragment"
         android:label="@string/title_stream"
         tools:layout="@layout/fragment_stream" >
         <action
@@ -20,7 +20,7 @@
 
     <fragment
         android:id="@+id/navigation_addtrip"
-        android:name="com.example.tripline.ui.addtrip.AddTripFragment"
+        android:name="com.example.tripline.fragments.AddTripFragment"
         android:label="@string/title_addtrip"
         tools:layout="@layout/fragment_addtrip" >
         <action
@@ -30,7 +30,7 @@
 
     <fragment
         android:id="@+id/navigation_addevent"
-        android:name="com.example.tripline.AddEventFragment"
+        android:name="com.example.tripline.fragments.AddEventFragment"
         android:label="@string/add_event"
         tools:layout="@layout/fragment_add_event" >
         <action
@@ -40,7 +40,7 @@
 
     <fragment
         android:id="@+id/navigation_tripdetails"
-        android:name="com.example.tripline.TripDetailsFragment"
+        android:name="com.example.tripline.fragments.TripDetailsFragment"
         android:label="@string/trip_details"
         tools:layout="@layout/fragment_trip_details" >
         <action
@@ -53,7 +53,7 @@
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.example.tripline.ui.profile.ProfileFragment"
+        android:name="com.example.tripline.fragments.ProfileFragment"
         android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile">
         <argument
@@ -75,7 +75,7 @@
 
     <fragment
         android:id="@+id/navigation_map"
-        android:name="com.example.tripline.MapFragment"
+        android:name="com.example.tripline.fragments.MapFragment"
         android:label="@string/map"
         tools:layout="@layout/fragment_map" >
         <action
@@ -85,7 +85,7 @@
 
     <fragment
         android:id="@+id/navigation_follower"
-        android:name="com.example.tripline.FollowerFragment"
+        android:name="com.example.tripline.fragments.FollowerFragment"
         android:label="Follower"
         tools:layout="@layout/fragment_follower" >
         <action
@@ -95,7 +95,7 @@
 
     <fragment
         android:id="@+id/navigation_following"
-        android:name="com.example.tripline.FollowingFragment"
+        android:name="com.example.tripline.fragments.FollowingFragment"
         android:label="Following"
         tools:layout="@layout/fragment_following" >
         <action
@@ -105,7 +105,7 @@
 
     <fragment
         android:id="@+id/navigation_search"
-        android:name="com.example.tripline.ui.search.SearchFragment"
+        android:name="com.example.tripline.fragments.SearchFragment"
         android:label="Search"
         tools:layout="@layout/fragment_search" >
         <action
